# docs/swagger.yaml
basePath: /
definitions:
  models.Avatar:
    properties:
      avatar_icon_url:
        type: string
      created_at:
        type: string
      id:
        type: string
      personality_traits:
        type: string
      prompt:
        type: string
      updated_at:
        type: string
      user_id:
        type: string
    type: object
  models.UserInfo:
    properties:
      created_at:
        type: string
      id:
        type: string
      is_mission_reward:
        description: ミッション報酬になっているかどうか
        type: boolean
      key:
        description: |-
          テキストの場合 Key: 項目名、Value: ユーザーが入力した値
          画像の場合 Key: 画像タイトル、Value: 画像のURL
        type: string
      type:
        $ref: '#/definitions/models.UserInfoType'
      updated_at:
        type: string
      user_id:
        type: string
      value:
        type: string
    type: object
  models.UserInfoType:
    enum:
    - text
    - image
    type: string
    x-enum-varnames:
    - UserInfoTypeText
    - UserInfoTypeImage
...
  requests.SendMatchMessageRequest:
    properties:
      content:
        example: こんにちは！マッチできて嬉しいです。
        type: string
    required:
    - content
    type: object
...
  requests.SendOnboardingMessageRequest:
    properties:
      content:
        example: 映画が好きです。
        type: string
    required:
    - content
    type: object
...
  requests.StartOnboardingRequest:
    type: object
...
  response.ErrorResponse:
    properties:
      error:
        example: invalid_token
        type: string
      message:
        example: Firebase IDトークンが不正または無効です
        type: string
    type: object
...
  response.GetMatchMessagesResponse:
    properties:
      limit:
        example: 20
        type: integer
      messages:
        items:
          $ref: '#/definitions/response.MatchMessage'
        type: array
      offset:
        example: 0
        type: integer
      total:
        example: 20
        type: integer
    type: object
...
  response.GetMatchesResponse:
    properties:
      matches:
        items:
          $ref: '#/definitions/response.Match'
        type: array
      total:
        example: 5
        type: integer
    type: object
...
  response.GetPartnersResponse:
    properties:
      limit:
        example: 20
        type: integer
      offset:
        example: 0
        type: integer
      partners:
        items:
          $ref: '#/definitions/response.Partner'
        type: array
      total:
        example: 50
        type: integer
    type: object
  response.GetSuggestionsResponse:
    properties:
      suggestions:
        items:
          $ref: '#/definitions/response.Suggestion'
        type: array
    type: object
...
  response.Match:
    properties:
      chat_id:
        example: 01ARZ3NDEKTSV4RRFFQ69G5FAV
        type: string
      final_score:
        example: 85
        type: integer
      id:
        example: 01ARZ3NDEKTSV4RRFFQ69G5FEV
        type: string
      matched_at:
        example: "2024-01-01T12:00:00Z"
        type: string
      partner_id:
        example: 01ARZ3NDEKTSV4RRFFQ69G5FBV
        type: string
      partner_image_url:
        example: https://example.com/images/profile2.jpg
        type: string
      partner_name:
        example: 佐藤花子
        type: string
      user_id:
        example: 01ARZ3NDEKTSV4RRFFQ69G5FAV
        type: string
    type: object
  response.MatchMessage:
    properties:
      content:
        example: こんにちは！マッチできて嬉しいです。
        type: string
      created_at:
        example: "2024-01-01T12:00:00Z"
        type: string
      id:
        example: 01ARZ3NDEKTSV4RRFFQ69G5FBV
        type: string
      match_id:
        example: 01ARZ3NDEKTSV4RRFFQ69G5FEV
        type: string
      sender_id:
        example: 01ARZ3NDEKTSV4RRFFQ69G5FAV
        type: string
    type: object
...
  response.OnboardingCompleteResponse:
    properties:
      avatar:
        $ref: '#/definitions/models.Avatar'
      user_infos:
        items:
          $ref: '#/definitions/models.UserInfo'
        type: array
    type: object
  response.Partner:
    properties:
      age:
        example: 24
        type: integer
      bio:
        example: よろしくお願いします！
        type: string
      display_name:
        example: 佐藤花子
        type: string
      gender:
        example: female
        type: string
      id:
        example: 01ARZ3NDEKTSV4RRFFQ69G5FBV
        type: string
      profile_image_url:
        example: https://example.com/images/profile2.jpg
        type: string
    type: object
...
  response.ReplyAssistResponse:
    properties:
      context:
        example: 初めてのメッセージに対する返信
        type: string
      suggested_replies:
        example:
        - '["ありがとうございます！こちらこそよろしくお願いします。"'
        - ' "こちらこそ！お話できるのを楽しみにしています。"]'
        items:
          type: string
        type: array
    type: object
...
  response.SendMatchMessageResponse:
    properties:
      message:
        $ref: '#/definitions/response.MatchMessage'
    type: object
...
  response.SendOnboardingMessageResponse:
    properties:
      message:
        example: メッセージを受け付けました
        type: string
    type: object
...
  response.StartOnboardingResponse:
    properties:
      chat_room_id:
        example: 01ARZ3NDEKTSV4RRFFQ69G5FAV
        type: string
      message:
        example: オンボーディングチャットを開始しました。Firestoreからメッセージを取得してください。
        type: string
    type: object
  response.Suggestion:
    properties:
      category:
        description: interests, hobbies, work, etc.
        example: interests
        type: string
      description:
        example: お互いの趣味について詳しく聞いてみましょう
        type: string
      examples:
        example:
        - '["どんな本が好きですか？"'
        - ' "最近読んだ本はありますか？"]'
        items:
          type: string
        type: array
      id:
        example: suggestion_123456
        type: string
      title:
        example: 共通の趣味について話す
        type: string
    type: object
...
info:
  contact: {}
  license:
    name: MIT
  title: Hackathon API
  version: "1.0"
paths:
...
  /matches:
    get:
      description: 自分がマッチングした相手の一覧を取得する
      responses:
        "200":
          description: マッチ一覧取得成功
          schema:
            $ref: '#/definitions/response.GetMatchesResponse'
        "401":
          description: 認証されていない、またはトークンが不正
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - Bearer: []
      summary: マッチ一覧取得
      tags:
      - matches
  /matches/{id}/messages:
    get:
      description: マッチした相手（本人）とのメッセージ履歴を取得する
      parameters:
      - description: マッチID
        in: path
        name: id
        required: true
        type: string
      - default: 20
        description: 取得件数
        in: query
        name: limit
        type: integer
      - default: 0
        description: オフセット
        in: query
        name: offset
        type: integer
      responses:
        "200":
          description: メッセージ履歴取得成功
          schema:
            $ref: '#/definitions/response.GetMatchMessagesResponse'
        "401":
          description: 認証されていない、またはトークンが不正
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "403":
          description: このマッチにアクセスする権限がない
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: マッチが見つからない
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - Bearer: []
      summary: 本人とのメッセージ履歴取得
      tags:
      - matches
    post:
      description: マッチした相手（本人）にメッセージを送信する
      parameters:
      - description: マッチID
        in: path
        name: id
        required: true
        type: string
      - description: メッセージ送信リクエスト
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/requests.SendMatchMessageRequest'
      responses:
        "200":
          description: メッセージ送信成功
          schema:
            $ref: '#/definitions/response.SendMatchMessageResponse'
        "400":
          description: リクエストが不正
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: 認証されていない、またはトークンが不正
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "403":
          description: このマッチにアクセスする権限がない
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: マッチが見つからない
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - Bearer: []
      summary: 本人とのメッセージ送信
      tags:
      - matches
  /matches/{id}/reply-assist:
    get:
      description: 相手からの最新メッセージに対する返信をAIが提案する（チャットルーム入室時や返信受信時に自動取得）
      parameters:
      - description: マッチID（ULID）
        in: path
        name: id
        required: true
        type: string
      - description: 返信対象のメッセージID（指定しない場合は最新メッセージ）
        in: query
        name: messageId
        type: string
      responses:
        "200":
          description: 返信アシスト取得成功
          schema:
            $ref: '#/definitions/response.ReplyAssistResponse'
        "401":
          description: 認証されていない、またはトークンが不正
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "403":
          description: このマッチにアクセスする権限がない
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: マッチが見つからない
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - Bearer: []
      summary: 返信アシスト
      tags:
      - matches
  /matches/{id}/suggestions:
    get:
      description: マッチした相手との会話を促進するための話題提案を取得する
      parameters:
      - description: マッチID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: 話題提案取得成功
          schema:
            $ref: '#/definitions/response.GetSuggestionsResponse'
        "401":
          description: 認証されていない、またはトークンが不正
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "403":
          description: このマッチにアクセスする権限がない
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: マッチが見つからない
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - Bearer: []
      summary: 話題提案取得
      tags:
      - matches
  /matches/candidates:
    get:
      description: マッチング前の相手候補の一覧を取得する（基本情報のみ）
      parameters:
      - default: 20
        description: 取得件数
        in: query
        name: limit
        type: integer
      - default: 0
        description: オフセット
        in: query
        name: offset
        type: integer
      responses:
        "200":
          description: 相手候補一覧取得成功
          schema:
            $ref: '#/definitions/response.GetPartnersResponse'
        "401":
          description: 認証されていない、またはトークンが不正
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - Bearer: []
      summary: 相手候補一覧取得
      tags:
      - matches
...
  /onboarding/chats/messages:
    post:
      description: オンボーディングチャットメッセージを送信する
      parameters:
      - description: オンボーディングチャットメッセージ送信リクエスト
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/requests.SendOnboardingMessageRequest'
      responses:
        "200":
          description: オンボーディングチャットメッセージ送信成功
          schema:
            $ref: '#/definitions/response.SendOnboardingMessageResponse'
        "400":
          description: リクエストが不正
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: 認証されていない、またはトークンが不正
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - Bearer: []
      summary: オンボーディングチャットメッセージ送信
      tags:
      - onboarding
  /onboarding/finish:
    post:
      description: オンボーディングチャットを完了する
      responses:
        "200":
          description: オンボーディングチャット完了成功
          schema:
            $ref: '#/definitions/response.OnboardingCompleteResponse'
        "400":
          description: リクエストが不正
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: 認証されていない、またはトークンが不正
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - Bearer: []
      summary: オンボーディングチャット完了
      tags:
      - onboarding
  /onboarding/start:
    post:
      description: オンボーディングチャットを開始する
      parameters:
      - description: オンボーディングチャット開始リクエスト
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/requests.StartOnboardingRequest'
      responses:
        "200":
          description: オンボーディングチャット開始成功
          schema:
            $ref: '#/definitions/response.StartOnboardingResponse'
        "400":
          description: リクエストが不正
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: 認証されていない、またはトークンが不正
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - Bearer: []
      summary: オンボーディングチャット開始
      tags:
      - onboarding
...
securityDefinitions:
  Bearer:
    description: 'Firebase IDトークンをBearerトークンとして送信してください。例: "Bearer eyJhbGciOiJSUzI1NiIs..."'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"